+++ src/manager/mod.rs
@@ -1,12 +1,12 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: MIT
 pub use evm::{gateway, EthManager, EthSubnetManager};
-pub use lotus::LotusSubnetManager;
+pub use fvm::LotusSubnetManager;
 pub use subnet::SubnetManager;
 
 pub use crate::lotus::message::ipc::SubnetInfo;
 
 pub mod evm;
 pub mod fevm;
-mod lotus;
+mod fvm;
 mod subnet;
